Improve reverse_string tool
